requirements.txt
orjson
PyMySQL
requests
websocket-client
cryptography
apscheduler
sortedcontainers
pyTelegramBotAPI
pytz
gspread
